requirements.txt
Flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
pymongo==4.6.0
PyJWT[crypto]==2.6.0
python-jose[cryptography]==3.3.0
requests==2.31.0
flask_cors==4.0.0
cryptography==41.0.7
redis==5.0.1
flask-caching==2.1.0
